Change GridCard and ListCard to ProductionGridCard and ProductionListCard to better reflect the components usage.
