+++ frontend/src/__tests__/components/ProductionListCard.test.tsx
@@ -4,7 +4,7 @@ import userEvent from '@testing-library/user-event'
 import type { ReactElement } from 'react'
 import { I18nextProvider } from 'react-i18next'
 import { MemoryRouter } from 'react-router-dom'
-import ListCard from '../../components/ListCard'
+import ProductionListCard from '../../components/ProductionListCard'
 import i18n from '../../i18n'
 import type { Event } from '../../types/Events'
 import type { Genre } from '../../types/Genres'
@@ -69,7 +69,7 @@ const renderListCard = (props: {
   const { production, selectedGenreIds = [], onGenreClick = jest.fn() } = props
 
   const ui: ReactElement = (
-    <ListCard
+    <ProductionListCard
       production={production}
       selectedGenreIds={selectedGenreIds}
       onGenreClick={onGenreClick}
@@ -93,7 +93,7 @@ afterEach(() => {
   void i18n.changeLanguage('nl')
 })
 
-describe('ListCard', () => {
+describe('ProductionListCard', () => {
   it('renders title, artist, full-card link, and poster image with title as alt text', () => {
     const production = baseProduction({
       media_gallery: {
